+++ bin/tx-prover/src/commands/mod.rs
@@ -1,5 +1,3 @@
-use std::{fs::File, io::Write};
-
 use clap::Parser;
 use figment::{
     providers::{Format, Toml},
@@ -9,9 +7,11 @@ use init::Init;
 use miden_tx_prover::PROVER_SERVICE_CONFIG_FILE_NAME;
 use proxy::StartProxy;
 use serde::{Deserialize, Serialize};
-use tracing::debug;
+use tracing::instrument;
 use update_workers::{AddWorkers, RemoveWorkers, UpdateWorkers};
 use worker::StartWorker;
+
+use crate::utils::MIDEN_TX_PROVER;
 
 pub mod init;
 pub mod proxy;
@@ -22,10 +22,8 @@ pub mod worker;
 ///
 /// It is stored in a TOML file, which will be created by the `init` command.
 /// It allows manual modification of the configuration file.
-#[derive(Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize)]
 pub struct ProxyConfig {
-    /// List of workers used by the proxy.
-    pub workers: Vec<WorkerConfig>,
     /// Host of the proxy.
     pub host: String,
     /// Port of the proxy.
@@ -49,7 +47,6 @@ pub struct ProxyConfig {
 impl Default for ProxyConfig {
     fn default() -> Self {
         Self {
-            workers: vec![WorkerConfig::new("0.0.0.0", 8083), WorkerConfig::new("0.0.0.0", 8084)],
             host: "0.0.0.0".into(),
             port: 8082,
             timeout_secs: 100,
@@ -76,55 +73,6 @@ impl ProxyConfig {
         Figment::from(Toml::file(config_path))
             .extract()
             .map_err(|err| format!("Failed to load {} config file: {err}", config_path.display()))
-    }
-
-    /// Saves the configuration to the config file
-    ///
-    /// This method will serialize the configuration to a TOML string and write it to the file with
-    /// the name defined at the [PROVER_SERVICE_CONFIG_FILE_NAME] constant in the current directory.
-    pub(crate) fn save_to_config_file(&self) -> Result<(), String> {
-        let mut current_dir = std::env::current_dir().map_err(|err| err.to_string())?;
-        current_dir.push(PROVER_SERVICE_CONFIG_FILE_NAME);
-        let config_path = current_dir.as_path();
-
-        let config_as_toml_string = toml::to_string_pretty(self)
-            .map_err(|err| format!("error formatting config: {err}"))?;
-
-        let mut file_handle = File::options()
-            .write(true)
-            .truncate(true)
-            .open(config_path)
-            .map_err(|err| format!("error opening the file: {err}"))?;
-
-        file_handle
-            .write(config_as_toml_string.as_bytes())
-            .map_err(|err| format!("error writing to file: {err}"))?;
-
-        debug!("Config updated successfully");
-
-        Ok(())
-    }
-
-    /// Updates the workers in the configuration with the new list.
-    pub(crate) fn set_workers(workers: Vec<WorkerConfig>) -> Result<(), String> {
-        let mut proxy_config = Self::load_config_from_file()?;
-
-        proxy_config.workers = workers;
-
-        proxy_config.save_to_config_file()
-    }
-}
-
-/// Configuration for a worker
-#[derive(Serialize, Deserialize)]
-pub struct WorkerConfig {
-    pub host: String,
-    pub port: u16,
-}
-
-impl WorkerConfig {
-    pub fn new(host: &str, port: u16) -> Self {
-        Self { host: host.into(), port }
     }
 }
 
@@ -156,18 +104,17 @@ pub enum Command {
     StartProxy(StartProxy),
     /// Adds workers to the proxy.
     ///
-    /// This method will make a request to the proxy defined in the config file to add workers. It
-    /// will update the configuration file with the new list of workers.
+    /// This method will make a request to the proxy defined in the config file to add workers.
     AddWorkers(AddWorkers),
     /// Removes workers from the proxy.
     ///
     /// This method will make a request to the proxy defined in the config file to remove workers.
-    /// It will update the configuration file with the new list of workers.
     RemoveWorkers(RemoveWorkers),
 }
 
 /// CLI entry point
 impl Cli {
+    #[instrument(target = MIDEN_TX_PROVER, name = "cli:execute", skip_all, ret(level = "info"), err)]
     pub async fn execute(&self) -> Result<(), String> {
         match &self.action {
             // For the `StartWorker` command, we need to create a new runtime and run the worker