+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changes
 
+- Added tracing to the `miden-tx-prover` CLI (#1014).
 - Added health check endpoints to the prover service (#1006).
 - Implemented serialization for `AccountHeader` (#996).
 - Updated Pingora crates to 0.4 and added polling time to the configuration file (#997).
@@ -18,6 +19,7 @@
 - [BREAKING] Added `miden::note::get_script_hash` procedure (#995).
 - [BREAKING] Refactor error messages in `miden-lib` and `miden-tx` and use `thiserror` 2.0 (#1005).
 - [BREAKING] Extend `AccountId` to two `Felt`s and require block hash in derivation (#982).
+- Removed workers list from the proxy configuration file (#1018).
 
 ## 0.6.2 (2024-11-20)